renamed push_T_easy.py to push_T.py
