+++ mani_skill/envs/tasks/digital_twins/__init__.py
@@ -1,1 +1,1 @@
-from .push_T_easy import PushTEnvEasy
+from .push_T import PushTEnvEasy
